perf: change the time variables type to "uint64"
perf: add type conversions in the "getWithdrawableAmount" function
test: add "assert" function for the "uint64" type
test: add "createDynamicArray" for the "uint64" type

+++ src/SablierV2.sol
@@ -225,8 +225,8 @@ abstract contract SablierV2 is ISablierV2 {
         address sender,
         address recipient,
         uint256 depositAmount,
-        uint256 startTime,
-        uint256 stopTime
+        uint64 startTime,
+        uint64 stopTime
     ) internal pure {
         // Checks: the sender is not the zero address.
         if (sender == address(0)) {
